added playertankcontroller using inheritance
