+++ Assets/Scripts/Tank/PlayerTankController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c56405596882beb4c949a9c1f3304513
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
